+++ tests/pre-commit.ps1
@@ -1,6 +1,6 @@
 Function RunTests {
+	$run_int_tests = "go clean -testcache; go test -v ."
 	$run_ext_tests = "go clean -testcache; go test -v .\tests\"
-	$run_int_tests = "go clean -testcache; go test -v .\voicemeeter\"
 
 	Invoke-Expression $run_ext_tests
 	Invoke-Expression $run_int_tests